Near 0.5.8... Added the decimals to the power

+++ trunk/src/qso.cpp
@@ -56,7 +56,7 @@ Qso::Qso(const QString& tqrz, int tnumb, QString& tmode){
 }
 
 
-Qso::Qso(const int tnumb, const QString& tqrz, const int trstrx, const int trsttx, const QString& tmode, const QString& tband, const int tpower){
+Qso::Qso(const int tnumb, const QString& tqrz, const int trstrx, const int trsttx, const QString& tmode, const QString& tband, const double tpower){
     qrz = tqrz;
     numb = tnumb;
     mode = tmode;
@@ -116,7 +116,7 @@ Qso::Qso(const int tnumb, const QDateTime tdateTime, const QString& tqrz, const 
 
 }
 
-Qso::Qso (const int tnumb, const QDateTime& tdateTime, const QString& tqrz, const int trstrx, const int trsttx, int tband, const QString& tmode, const int tpower, const QDate& tqslSen,  const QDate& tqslRec, const QString& tcomment){
+Qso::Qso (const int tnumb, const QDateTime& tdateTime, const QString& tqrz, const int trstrx, const int trsttx, int tband, const QString& tmode, const double tpower, const QDate& tqslSen,  const QDate& tqslRec, const QString& tcomment){
   qrz = checkisNotAdif(tqrz);
   numb = tnumb;
   dateTime = tdateTime;
@@ -411,12 +411,18 @@ QString Qso::getMode() const{
 //qDebug() << "QSO::getMode: " << mode << endl;
     return mode;}
 
-void Qso::setPower(const QString& tpower){
-    power = checkisNotAdif(tpower);
+void Qso::setPower(const QString& tpower)
+{
+  //qDebug() << "QSO::setPower: " << tpower<< endl;
+  power = checkisNotAdif(tpower);
 }
 
-QString Qso::getPower() const{
-    return power;}
+QString Qso::getPower() const
+{
+    //qDebug() << "QSO::getPower: " << power<< endl;
+    return power;
+  
+}
 
 void Qso::setrxPower(const QString& tpower){
 //	cout << "Qso::setrxPower: " << rx_power << endl;
